inet: remove inheritance to PlainVehicle & move mobility module to the end of the submodule list to avoid a UAF in other modules upon module desctruction
OMNet++ initializes the submodules of the module's super type first and
adds them to the list of the module's submodules, then does the same for
the module's submodules. OMNet++ destroys submodules in the order they
are listed in this list, thus destroying the super types' submodules
first. This leads to a UAF if a submodule needs to access the mobility
module in the destructor, like VanetNic.