Middleware: fix initialization of multi-stage services
Problem: Services with multiple stages are initilialized twice.

Calling createScheduleInit() will initialize all stages of a module.
This is done in stage 1 from the initialize function of the Middleware.

If a services has 2 or more stages, all stages greater or equal to 1
will be initialized twice, as later stages are handled by OMNeT++ again.

Change-Id: Ia7a7df456fce0c55938d9cb130a3e681c5e40945