Destruction of vehicles on VAR_TELEPORT_STARTING_VEHICLES_IDS TraCI signal
In SUMO vehicles get teleported when they are either crashing or when
waiting to long (https://sumo.dlr.de/wiki/Simulation/Why_Vehicles_are_teleporting).
SUMO can also be condigured to destroy both involved vehicles on a crash
(using the '--collision.action' paramter).
But in certain cases it is necessary to destroy vehicles from Artery so
that it is possible to record the conditions under which the crash
occured. This behavior is also implemented in the TraCIScenarioManager
of Veins.

The commit includes the possiblity to set the boolean 'destroyOnCrash' parameter
for the 'traci.core' module, which then controls the destruction of
crashing vehicles.