traci: destruct vehicles on teleportation
In SUMO vehicles get teleported when they are either crashing or when
waiting to long (https://sumo.dlr.de/wiki/Simulation/Why_Vehicles_are_teleporting).
SUMO can also be condigured to destroy both involved vehicles on a crash
(using the '--collision.action' paramter).
But in certain cases it is necessary to destroy vehicles from Artery so
that it is possible to record the conditions under which the crash
occured. This behavior is also implemented in the TraCIScenarioManager
of Veins.

The commit adds the 'destroyVehiclesOnCrash' flag for the BasicNodeManager,
which then controls the destruction of crashing vehicles.

Change-Id: Id01b04e385e06baa396d18c34fcf803a268e019a