Improves performance of channelBusyRatio
For calculation of the CBR a circular_buffer, storing a bool for
every symbol in the CBR interval (T_{CBR}=100ms) was used.
But counting and inserting a high amount of elements was inefficient.

This commit replaces the circular_buffer with a list of tuples. Each
tuple stores the amount of symbols and if the medium was busy for the
amount of symbols, reducing the number of elements needed.