Fix ItsG5BaseService multi-stage initialization
Just overriding the staged initialize method is not enough, the returned
number of init stages must also be adjusted to the number of required
stages. Additionally update the code to allow derived classes to use
single-stage initialization easily, for these classes it is enough to
just override the single-stage initialization method, it is not required
anymore to overrided the multi-stage initialization method.