envmod: rework of sensor abstraction and configuration
The previous RadarSensor is now the more general FovSensor.
The new RadarSensor specifies the sensor category and LoS checks for a
FovSensor.

The new OminiscientSensor is a FovSensor with a 360 dregree view and of
category 'omniscient' that does not perform LoS checks.

Additionally, sensors can be named to distinguish them from each other when they are of the same type. The name can be set using the XML attribute `name`.
To determine objects detected by a sensors with given name, the function `filterBySensorName` of the LocalEnvironmentModel can be used.
The usage is shown in the artery scenario.

EnvmodPrinter now also prints known objects by sensor name.