Add Glossary Feature (#307)
* Add abbreviations list and integrate into docs

Added a new file (`abbreviations.md`) which contains a comprehensive list of abbreviations used throughout the project documentation. This was particularly necessary to ensure that all documentation readers are familiar with the specific language and terminologies used herein. The abbreviations list was also integrated into the project markdown extensions for automatic appending in documents, improving readability and understandability of the project documentation.

* Fixed Mistake

* Removed ESP from Abbreviations.md

* Fixed Typo